# backend/handlers/project.go
package handlers

import (
	"net/http"
	"strings"

	"github.com/gin-gonic/gin"
	"github.com/tonk/warmdesk/database"
	"github.com/tonk/warmdesk/middleware"
	"github.com/tonk/warmdesk/models"
	"github.com/tonk/warmdesk/services"
)

// labelPalette is cycled through when assigning colors to default labels.
var labelPalette = []string{
	"#ef4444", // red
	"#3b82f6", // blue
	"#8b5cf6", // purple
	"#10b981", // green
	"#f59e0b", // amber
	"#06b6d4", // cyan
	"#ec4899", // pink
	"#84cc16", // lime
}

type labelDef struct {
	Name  string
	Color string
}

// getDefaultLabelDefs reads the configured initial label names from system settings
// and pairs them with colors from the palette.
func getDefaultLabelDefs() []labelDef {
	all := loadAllSettings()
	raw := all[settingDefaultLabels]
	var defs []labelDef
	for i, line := range strings.Split(raw, "\n") {
		name := strings.TrimSpace(line)
		if name == "" {
			continue
		}
		color := labelPalette[i%len(labelPalette)]
		defs = append(defs, labelDef{Name: name, Color: color})
	}
	return defs
}

// getDefaultColumnNames reads the configured initial column names from system settings.
func getDefaultColumnNames() []string {
	all := loadAllSettings()
	raw := all[settingDefaultColumns]
	var names []string
	for _, line := range strings.Split(raw, "\n") {
		name := strings.TrimSpace(line)
		if name != "" {
			names = append(names, name)
		}
	}
	if len(names) == 0 {
		return []string{"Backlog"}
	}
	return names
}

type ProjectListItem struct {
	models.Project
	OpenCardCount int64 `json:"open_card_count"`
}

func projectsWithCounts(projects []models.Project) []ProjectListItem {
	result := make([]ProjectListItem, len(projects))
	for i, p := range projects {
		var count int64
		database.DB.Model(&models.Card{}).Where("project_id = ? AND closed = false", p.ID).Count(&count)
		result[i] = ProjectListItem{Project: p, OpenCardCount: count}
	}
	return result
}

// ListProjects godoc
// @Summary      List projects accessible to the current user
// @Tags         projects
// @Produce      json
// @Security     BearerAuth
// @Success      200 {array}  models.Project
// @Router       /projects [get]
func ListProjects(c *gin.Context) {
	userID := middleware.GetUserID(c)
	globalRole := middleware.GetGlobalRole(c)

	// Admins and global viewers see all non-deleted projects
	if globalRole == "admin" || globalRole == "viewer" {
		var projects []models.Project
		database.DB.Where("deleted_at IS NULL").Find(&projects)
		c.JSON(http.StatusOK, projectsWithCounts(projects))
		return
	}

	var members []models.ProjectMember
	database.DB.Preload("Project").Where("user_id = ?", userID).Find(&members)

	projects := make([]models.Project, 0, len(members))
	for _, m := range members {
		if m.Project.DeletedAt.Valid {
			continue
		}
		projects = append(projects, m.Project)
	}
	c.JSON(http.StatusOK, projectsWithCounts(projects))
}

// CreateProject godoc
// @Summary      Create a new project
// @Tags         projects
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        body body map[string]string true "Project name, description, color"
// @Success      201 {object} models.Project
// @Failure      400 {object} map[string]string
// @Failure      403 {object} map[string]string
// @Router       /projects [post]
func CreateProject(c *gin.Context) {
	userID := middleware.GetUserID(c)
	if middleware.GetGlobalRole(c) == "viewer" {
		c.JSON(http.StatusForbidden, gin.H{"error": "viewers cannot create projects"})
		return
	}
	var req struct {
		Name        string `json:"name" binding:"required,min=1,max=200"`
		Description string `json:"description"`
		Color       string `json:"color"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	project := models.Project{
		Name:        req.Name,
		Description: req.Description,
		Color:       req.Color,
		Slug:        services.GenerateSlug(req.Name),
		KeyPrefix:   services.GenerateKeyPrefix(req.Name),
		CreatedByID: userID,
	}

	if err := database.DB.Create(&project).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "internal error"})
		return
	}

	// Creator becomes owner
	member := models.ProjectMember{
		ProjectID: project.ID,
		UserID:    userID,
		Role:      "owner",
		InvitedBy: userID,
	}
	database.DB.Create(&member)

	// Default columns from system settings
	for i, name := range getDefaultColumnNames() {
		database.DB.Create(&models.Column{ProjectID: project.ID, Name: name, Position: float64((i + 1) * 1000)})
	}

	// Default labels from system settings
	for _, def := range getDefaultLabelDefs() {
		database.DB.Create(&models.Label{ProjectID: project.ID, Name: def.Name, Color: def.Color})
	}

	c.JSON(http.StatusCreated, project)
}

// GetProject godoc
// @Summary      Get a project with its columns, cards, labels and members
// @Tags         projects
// @Produce      json
// @Security     BearerAuth
// @Param        projectSlug path string true "Project slug"
// @Success      200 {object} models.Project
// @Failure      403 {object} map[string]string
// @Failure      404 {object} map[string]string
// @Router       /projects/{projectSlug} [get]
func GetProject(c *gin.Context) {
	userID := middleware.GetUserID(c)
	slug := c.Param("projectSlug")

	project, err := services.GetProjectBySlug(slug)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "project not found"})
		return
	}

	if err := services.RequireProjectRole(project.ID, userID, middleware.GetGlobalRole(c), "viewer"); err != nil {
		c.JSON(http.StatusForbidden, gin.H{"error": "forbidden"})
		return
	}

	database.DB.Preload("Columns").Preload("Columns.Cards").Preload("Columns.Cards.Assignee").Preload("Columns.Cards.Labels").Preload("Columns.Cards.Tags").Preload("Labels").Preload("Members.User").First(project, project.ID)

	c.JSON(http.StatusOK, project)
}

// UpdateProject godoc
// @Summary      Update a project (owner/admin only)
// @Tags         projects
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        projectSlug path string true "Project slug"
// @Param        body body map[string]interface{} true "Fields to update"
// @Success      200 {object} models.Project
// @Failure      403 {object} map[string]string
// @Failure      404 {object} map[string]string
// @Router       /projects/{projectSlug} [put]
func UpdateProject(c *gin.Context) {
	userID := middleware.GetUserID(c)
	slug := c.Param("projectSlug")

	project, err := services.GetProjectBySlug(slug)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "project not found"})
		return
	}

	if err := services.RequireProjectRole(project.ID, userID, middleware.GetGlobalRole(c), "owner"); err != nil {
		c.JSON(http.StatusForbidden, gin.H{"error": "forbidden"})
		return
	}

	var req struct {
		Name        string `json:"name"`
		Description string `json:"description"`
		Color       string `json:"color"`
		IsArchived  *bool  `json:"is_archived"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	updates := map[string]interface{}{}
	if req.Name != "" {
		updates["name"] = req.Name
	}
	if req.Description != "" {
		updates["description"] = req.Description
	}
	if req.Color != "" {
		updates["color"] = req.Color
	}
	if req.IsArchived != nil {
		updates["is_archived"] = *req.IsArchived
	}

	database.DB.Model(project).Updates(updates)
	database.DB.First(project, project.ID)
	c.JSON(http.StatusOK, project)
}

// DeleteProject godoc
// @Summary      Delete a project (owner/admin only)
// @Tags         projects
// @Produce      json
// @Security     BearerAuth
// @Param        projectSlug path string true "Project slug"
// @Success      204
// @Failure      403 {object} map[string]string
// @Failure      404 {object} map[string]string
// @Router       /projects/{projectSlug} [delete]
func DeleteProject(c *gin.Context) {
	userID := middleware.GetUserID(c)
	slug := c.Param("projectSlug")

	project, err := services.GetProjectBySlug(slug)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "project not found"})
		return
	}

	globalRole := middleware.GetGlobalRole(c)
	if globalRole != "admin" {
		if err := services.RequireProjectRole(project.ID, userID, globalRole, "owner"); err != nil {
			c.JSON(http.StatusForbidden, gin.H{"error": "forbidden"})
			return
		}
	}

	database.DB.Delete(project)
	c.JSON(http.StatusOK, gin.H{"message": "deleted"})
}